added full unicode support (tags + filename)
git-svn-id: https://intra.ist.tugraz.at/svn/tagstore@438 20732adb-bd90-46d1-8e52-e978120fd635

+++ research_platform/tagstore/tsos/linux.py
@@ -50,8 +50,9 @@ class FileSystem():
             os.unlink(link)
         ## delete folder if empty
         parent_path = os.path.dirname(link)
-        if len(os.listdir(parent_path)) == 0: #empty
-            os.rmdir(parent_path)        
+        if os.path.exists(parent_path):             #existing
+            if len(os.listdir(parent_path)) == 0:   #empty
+                os.rmdir(parent_path)        
         
     def inode_shortage(self, file_path, threshold_pct):
         """
