+++ research_platform/tagstore/tsos/windows.py
@@ -49,9 +49,10 @@ class FileSystem():
             os.remove(file_path)
 
         ## delete folder if empty
-        parent_path = os.path.dirname(link)         
-        if len(os.listdir(parent_path)) == 0: #empty
-            os.rmdir(parent_path)        
+        parent_path = os.path.dirname(unicode(link))
+        if os.path.exists(parent_path):             #existing
+            if len(os.listdir(parent_path)) == 0:   #empty
+                os.rmdir(parent_path)        
         
     def inode_shortage(self, file_path, threshold_pct):
         """
